# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for managing mail system",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/user": {
            "post": {
                "description": "Create a new order with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Create a new UserAuthRequest",
                "parameters": [
                    {
                        "description": "The input todo struct",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UserAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserAuthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.UserAuthRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "pass": {
                    "type": "string"
                }
            }
        },
        "handlers.UserAuthResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "integer"
                }
            }
        }
    }
}